refactor(store): split databases, schemas and agents out of action.ts (#711)
action.ts 2,769 -> 1,761. Same mechanics as the first cut: byte-for-byte
moves with the doc comment above each declaration, action.ts still the
barrel, no consumer import changes.

Extraction order is topological, and that is load-bearing. Moving a
module before the ones it depends on would leave it importing from
action.ts -- which re-exports it -- so the first incremental cut would
have introduced the cycle the whole design avoids. databases, schemas
and agents depend only on shared, so they are safe to move now; forms,
scopes, fields and views wait for the modules they reach into.

Three symbols gain `export` because concerns still in action.ts reach
them: sortAndRemoveDupSchemas (fetchScopes), addNsfDesign (pullForms),
isActiveAgent (processViewsAgents). All three imports disappear in the
final cut, when those callers move to their own modules.

25 further bindings pruned from action.ts's import block.

lint 0 · build 0 · 94 files / 1138 tests, unchanged
